Extract BucketListItemForm and refactor
